Use a dylib for linking to typst during development
