+++ backend/Cargo.toml
@@ -1,6 +1,13 @@
 [workspace]
 resolver = "3"
-members = [".", "apportionment", "pdf_gen"]
+members = [
+    ".",
+    "apportionment",
+    "pdf_gen/types",
+    "pdf_gen/impl",
+    "pdf_gen/dylib",
+    "pdf_gen/wrapper",
+]
 
 [workspace.package]
 version = "1.1.0"
@@ -9,6 +16,10 @@ license = "EUPL-1.2"
 rust-version = "1.89"
 
 [workspace.dependencies]
+pdf_gen = { path = "./pdf_gen/wrapper" }
+pdf_gen_impl = { path = "./pdf_gen/impl" }
+pdf_gen_types = { path = "./pdf_gen/types" }
+pdf_gen_dylib = { path = "./pdf_gen/dylib" }
 async_zip = { version = "0.0.18", features = ["tokio", "deflate", "chrono"] }
 axum = { version = "0.8", features = ["macros", "tracing", "tokio"] }
 axum-extra = { version = "0.12", default-features = false, features = [
@@ -34,10 +45,10 @@ rust-version.workspace = true
 default-run = "abacus"
 
 [features]
-default = ["openapi", "dev-database"]
+default = ["openapi", "dev-database", "pdf_gen/dev"]
 openapi = ["dep:utoipa-swagger-ui"]
 memory-serve = ["dep:memory-serve"]
-embed-typst = ["pdf_gen/embed-typst"]
+embed-typst = ["pdf_gen/static"]
 dev-database = []
 airgap-detection = []
 storybook = ["dep:memory-serve"]
@@ -52,13 +63,12 @@ required-features = ["dev-database"]
 
 [[bin]]
 name = "gen-pdf"
-required-features = ["embed-typst"]
 
 # Note: if you add a dependency here, please update the README.md for the backend as well.
 
 [dependencies]
 apportionment = { path = "./apportionment" }
-pdf_gen = { path = "./pdf_gen" }
+pdf_gen.workspace = true
 axum.workspace = true
 axum-extra.workspace = true
 chrono.workspace = true
