+++ backend/pdf_gen/types/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "pdf_gen"
+name = "pdf_gen_types"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
@@ -8,21 +8,14 @@ rust-version.workspace = true
 [lints]
 workspace = true
 
-[features]
-embed-typst = ["dep:typst", "dep:typst-pdf"]
-
 [dependencies]
 async_zip.workspace = true
 axum-extra.workspace = true
 axum.workspace = true
-chrono.workspace = true
-rand.workspace = true
-strum.workspace = true
-tokio-util = { version = "0.7.18", features = ["io"] }
 tokio.workspace = true
-tracing.workspace = true
-typst = { version = "0.14.2", optional = true }
-typst-pdf = { version = "0.14.2", optional = true }
+tokio-util = { version = "0.7.18", features = ["io"] }
+chrono.workspace = true
+strum.workspace = true
 
 [dev-dependencies]
 http-body-util.workspace = true
