+++ backend/pdf_gen/wrapper/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "pdf_gen"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+rust-version.workspace = true
+
+[features]
+default = []
+static = ["dep:pdf_gen_impl"]
+dev = ["dep:pdf_gen_dylib", "dep:libloading"]
+
+[dependencies]
+pdf_gen_impl = { workspace = true, optional = true }
+pdf_gen_types.workspace = true
+pdf_gen_dylib = { workspace = true, optional = true }
+libloading = { version = "0.9.0", optional = true }
+tokio.workspace = true
+tracing.workspace = true
