+++ backend/src/api/document.rs
@@ -116,7 +116,7 @@ async fn election_download_n_10_2(
     let (zip_response, zip_writer) = ZipResponse::new(&zip_filename);
 
     tokio::spawn(async move {
-        if let Err(e) = generate_pdfs(&models, zip_writer).await {
+        if let Err(e) = generate_pdfs(models, zip_writer).await {
             error!("Failed to generate PDFs: {e:?}");
         }
     });
@@ -205,7 +205,7 @@ async fn election_download_na_31_2_bijlage1(
     let (zip_response, zip_writer) = ZipResponse::new(&zip_filename);
 
     tokio::spawn(async move {
-        if let Err(e) = generate_pdfs(&models, zip_writer).await {
+        if let Err(e) = generate_pdfs(models, zip_writer).await {
             error!("Failed to generate PDFs: {e:?}");
         }
     });
@@ -259,7 +259,7 @@ async fn election_download_na_31_2_inlegvel(
     }
     .to_pdf_file_model(name.clone());
 
-    let content = generate_pdf(&input).await?;
+    let content = generate_pdf(input).await?;
 
     Ok(Attachment::new(content.buffer)
         .filename(&name)
