+++ backend/src/bin/gen-pdf.rs
@@ -67,7 +67,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         let file_name = format!("{path}/{}.pdf", model.as_model_name());
         let file_model = PdfFileModel::new(file_name.clone(), model);
 
-        let pdf = match generate_pdf(&file_model).await {
+        let pdf = match generate_pdf(file_model).await {
             Ok(pdf) => pdf,
             Err(e) => {
                 eprintln!("Error generating PDF {file_name}: {e:?}");
